tests/wpt/web-platform-tests/webrtc-insertable-streams/RTCPeerConnection-insertable-streams.https.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection Insertable Streams</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../webrtc/RTCPeerConnection-helper.js"></script>
<script src="./RTCPeerConnection-insertable-streams.js"></script>
</head>
<body>
<script>
async function testVideoFlow(t, negotiationFunction) {
  const caller = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => callee.close());

  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const videoTrack = stream.getVideoTracks()[0];
  t.add_cleanup(() => videoTrack.stop());

  const videoSender = caller.addTrack(videoTrack)
  const senderStreams = videoSender.createEncodedStreams();
  const senderReader = senderStreams.readable.getReader();
  const senderWriter = senderStreams.writable.getWriter();

  const frameInfos = [];
  const numFramesPassthrough = 5;
  const numFramesReplaceData = 5;
  const numFramesModifyData = 5;
  const numFramesToSend = numFramesPassthrough + numFramesReplaceData + numFramesModifyData;

  const ontrackPromise = new Promise(resolve => {
    callee.ontrack = t.step_func(() => {
      const videoReceiver = callee.getReceivers().find(r => r.track.kind === 'video');
      assert_true(videoReceiver !== undefined);

      const receiverStreams =
          videoReceiver.createEncodedStreams();
      const receiverReader = receiverStreams.readable.getReader();
      const receiverWriter = receiverStreams.writable.getWriter();

      const maxFramesToReceive = numFramesToSend;
      let numVerifiedFrames = 0;
      for (let i = 0; i < maxFramesToReceive; i++) {
        receiverReader.read().then(t.step_func(result => {
          verifyNonstandardAdditionalDataIfPresent(result.value);
          if (frameInfos[numVerifiedFrames] &&
              areFrameInfosEqual(result.value, frameInfos[numVerifiedFrames])) {
            numVerifiedFrames++;
          } else {
            // Receiving unexpected frames is an indication that
            // frames are not passed correctly between sender and receiver.
            assert_unreached("Incorrect frame received");
          }

          if (numVerifiedFrames == numFramesToSend)
            resolve();
        }));
      }
    });
  });

  exchangeIceCandidates(caller, callee);
  await negotiationFunction(caller, callee);

  // Pass frames as they come from the encoder.
  for (let i = 0; i < numFramesPassthrough; i++) {
    const result = await senderReader.read();
    const metadata = result.value.getMetadata();
    assert_true(containsVideoMetadata(metadata));
    verifyNonstandardAdditionalDataIfPresent(result.value);
    frameInfos.push({
      timestamp: result.value.timestamp,
      type: result.value.type,
      data: result.value.data,
      metadata: metadata,
      getMetadata() { return this.metadata; }
    });
    senderWriter.write(result.value);
  }

  // Replace frame data with arbitrary buffers.
  for (let i = 0; i < numFramesReplaceData; i++) {
    const result = await senderReader.read();
    const metadata = result.value.getMetadata();
    assert_true(containsVideoMetadata(metadata));
    const buffer = new ArrayBuffer(100);
    const int8View = new Int8Array(buffer);
    int8View.fill(i);

    result.value.data = buffer;
    frameInfos.push({
      timestamp: result.value.timestamp,
      type: result.value.type,
      data: result.value.data,
      metadata: metadata,
      getMetadata() { return this.metadata; }
    });
    senderWriter.write(result.value);
  }

  // Modify frame data.
  for (let i = 0; i < numFramesReplaceData; i++) {
    const result = await senderReader.read();
    const metadata = result.value.getMetadata();
    assert_true(containsVideoMetadata(metadata));
    const int8View = new Int8Array(result.value.data);
    int8View.fill(i);

    frameInfos.push({
      timestamp: result.value.timestamp,
      type: result.value.type,
      data: result.value.data,
      metadata: metadata,
      getMetadata() { return this.metadata; }
    });
    senderWriter.write(result.value);
  }

  return ontrackPromise;
}

promise_test(async t => {
  return testVideoFlow(t, exchangeOfferAnswer);
}, 'Frames flow correctly using insertable streams');

promise_test(async t => {
  return testVideoFlow(t, exchangeOfferAnswerReverse);
}, 'Frames flow correctly using insertable streams when receiver starts negotiation');

promise_test(async t => {
  const caller = new RTCPeerConnection();
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection();
  t.add_cleanup(() => callee.close());
  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const videoTrack = stream.getVideoTracks()[0];
  t.add_cleanup(() => videoTrack.stop());

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);

  const videoSender = caller.addTrack(videoTrack);
  assert_throws_dom("InvalidStateError", () => videoSender.createEncodedStreams());
}, 'RTCRtpSender.createEncodedStream() throws if not requested in PC configuration');

promise_test(async t => {
  const caller = new RTCPeerConnection();
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection();
  t.add_cleanup(() => callee.close());
  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const videoTrack = stream.getVideoTracks()[0];
  t.add_cleanup(() => videoTrack.stop());

  const videoSender = caller.addTrack(videoTrack);
  const ontrackPromise = new Promise(resolve => {
    callee.ontrack = t.step_func(() => {
      const videoReceiver = callee.getReceivers().find(r => r.track.kind === 'video');
      assert_true(videoReceiver !== undefined);
      assert_throws_dom("InvalidStateError", () => videoReceiver.createEncodedStreams());
      resolve();
    });
  });

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);
  return ontrackPromise;
}, 'RTCRtpReceiver.createEncodedStream() throws if not requested in PC configuration');

promise_test(async t => {
  const caller = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection();
  t.add_cleanup(() => callee.close());

  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const videoTrack = stream.getVideoTracks()[0];
  t.add_cleanup(() => videoTrack.stop());

  const videoSender = caller.addTrack(videoTrack)
  const senderStreams = videoSender.createEncodedStreams();

  const senderWorker = new Worker('RTCPeerConnection-sender-worker-single-frame.js')
  senderWorker.postMessage(
    {readableStream: senderStreams.readable},
    [senderStreams.readable]);

  let expectedFrameData = null;
  let verifiedFrameData = false;
  let numVerifiedFrames = 0;
  const onmessagePromise = new Promise(resolve => {
    senderWorker.onmessage = t.step_func(message => {
      if (!(message.data instanceof RTCEncodedVideoFrame)) {
        // This is the first message sent from the Worker to the test.
        // It contains an object (not an RTCEncodedVideoFrame) with the same
        // fields as the RTCEncodedVideoFrame to be sent in follow-up messages.
        // These serve as expected values to validate that the
        // RTCEncodedVideoFrame is sent correctly back to the test in the next
        // message.
        expectedFrameData = message.data;
      } else {
        // This is the frame sent by the Worker after reading it from the
        // readable stream. The Worker sends it twice after sending the
        // verification message.
        assert_equals(message.data.type, expectedFrameData.type);
        assert_equals(message.data.timestamp, expectedFrameData.timestamp);
        assert_true(areArrayBuffersEqual(message.data.data, expectedFrameData.data));
        assert_equals(message.data.getMetadata().synchronizationSource, expectedFrameData.metadata.synchronizationSource);
        if (++numVerifiedFrames == 2)
          resolve();
      }
    });
  });

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);

  return onmessagePromise;
}, 'RTCRtpSender readable stream transferred to a Worker and the Worker sends an RTCEncodedVideoFrame back');

async function testNormalDataFlowWithInsertableStreamsForOtherKind(t, normalKind) {
  const flowAudio = (normalKind == 'audio');
  const pcConfig = flowAudio
    ? {forceEncodedVideoInsertableStreams:true}
    : {forceEncodedAudioInsertableStreams:true};
  const caller = new RTCPeerConnection(pcConfig);
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection(pcConfig);
  t.add_cleanup(() => callee.close());

  const stream = await navigator.mediaDevices.getUserMedia(
    {audio: flowAudio, video: !flowAudio});
  const track = stream.getTracks()[0];
  t.add_cleanup(() => track.stop());

  const sender = caller.addTrack(track, stream);
  const ontrackPromise = new Promise((resolve,reject) => {
    callee.ontrack =  t.step_func(async e => {
      let numGetStats = 0;
      const intervalId = window.setInterval(async ()=> {
        let statsReport = await callee.getStats();
        statsReport.forEach(report => {
          if (report.type == 'inbound-rtp' &&
              report.kind == normalKind &&
              report.bytesReceived > 0) {
            clearInterval(intervalId);
            resolve();
            return;
          }
        });
        if (++numGetStats >= 60) {
          clearInterval(intervalId);
          reject('No bytes received.');
          return;
        }
      }, 50);
    });
  });

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);

  return ontrackPromise;
}

promise_test(t => {
  return testNormalDataFlowWithInsertableStreamsForOtherKind(t, 'audio');
}, 'Legacy: Audio flows when insertable streams is enabled for video and disabled for audio.');

promise_test(t => {
  return testNormalDataFlowWithInsertableStreamsForOtherKind(t, 'video');
}, 'Legacy: Video flows when insertable streams is enabled for audio and disabled for audio.');

promise_test(async t => {
  const caller = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => callee.close());

  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const track = stream.getTracks()[0];
  t.add_cleanup(() => track.stop());

  const sender = caller.addTrack(track)
  const senderStreams = sender.createEncodedStreams();
  const senderReader = senderStreams.readable.getReader();
  const senderWriter = senderStreams.writable.getWriter();
  const numFramesToSend = 20;

  const ontrackPromise = new Promise((resolve, reject) => {
    callee.ontrack = async e => {
      const receiverStreams = e.receiver.createEncodedStreams();
      const receiverReader = receiverStreams.readable.getReader();

      let numReceivedKeyFrames = 0;
      let numReceivedDeltaFrames = 0;
      for (let i = 0; i < numFramesToSend; i++) {
        const result = await receiverReader.read();
        if (result.value.type == 'key')
          numReceivedKeyFrames++;
        else if (result.value.type == 'delta')
          numReceivedDeltaFrames++;

        if (numReceivedKeyFrames > 0 && numReceivedDeltaFrames > 0)
          resolve();
        else if (numReceivedKeyFrames + numReceivedDeltaFrames >= numFramesToSend)
          reject();
      }
    }
  });

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);

  let numSentKeyFrames = 0;
  let numSentDeltaFrames = 0;
  // Pass frames as they come from the encoder.
  for (let i = 0; i < numFramesToSend; i++) {
    const result = await senderReader.read();
    verifyNonstandardAdditionalDataIfPresent(result.value);
    if (result.value.type == 'key') {
      numSentKeyFrames++;
    } else {
      numSentDeltaFrames++;
    }

    senderWriter.write(result.value);
  }

  assert_greater_than(numSentKeyFrames, 0);
  assert_greater_than(numSentDeltaFrames, 0);

  return ontrackPromise;
}, 'Key and Delta frames are sent and received');

promise_test(async t => {
  const caller = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection();
  t.add_cleanup(() => callee.close());

  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const track = stream.getTracks()[0];
  t.add_cleanup(() => track.stop());

  const sender = caller.addTrack(track)
  const streams = sender.createEncodedStreams();
  const transformer = new TransformStream({
    transform(frame, controller) {
      // Inserting the same frame twice will result in failure since the frame
      // will be neutered after the first insertion is processed.
      controller.enqueue(frame);
      controller.enqueue(frame);
    }
  });

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);

  await promise_rejects_dom(
      t, 'OperationError',
      streams.readable.pipeThrough(transformer).pipeTo(streams.writable));
}, 'Enqueuing the same frame twice fails');

promise_test(async t => {
  const caller = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => caller.close());
  const callee = new RTCPeerConnection({encodedInsertableStreams:true});
  t.add_cleanup(() => callee.close());

  const stream = await navigator.mediaDevices.getUserMedia({video:true});
  const videoTrack = stream.getVideoTracks()[0];
  t.add_cleanup(() => videoTrack.stop());

  const videoSender = caller.addTrack(videoTrack)
  const senderStreams = videoSender.createEncodedStreams();
  const senderReader = senderStreams.readable.getReader();
  const senderWriter = senderStreams.writable.getWriter();

  let sentFrameInfo = null;
  const ontrackPromise = new Promise(resolve => {
    callee.ontrack = t.step_func(async e => {
      const videoReceiver = e.receiver;

      const receiverStreams = videoReceiver.createEncodedStreams();
      const receiverReader = receiverStreams.readable.getReader();
      receiverReader.read().then(t.step_func(receivedResult => {
        assert_true(areFrameInfosEqual(receivedResult.value, sentFrameInfo));
        verifyNonstandardAdditionalDataIfPresent(receivedResult.value);
        resolve();
      }));
    });
  });

  exchangeIceCandidates(caller, callee);
  await exchangeOfferAnswer(caller, callee);

  // Pass frames as they come from the encoder.
  const result = await senderReader.read();
  verifyNonstandardAdditionalDataIfPresent(result.value);
  sentFrameInfo = {
    timestamp: result.value.timestamp,
    type: result.value.type,
    data: result.value.data,
    metadata: result.value.getMetadata(),
    getMetadata() { return this.metadata; }
  }
  senderWriter.write(result.value);

  return ontrackPromise;
}, 'Legacy API works');

</script>
</body>
</html>
